Fix duplicate game location issue
It turns out that xsi:type attributes of <GameLocation> tags aren't
unique; there are three <GameLocation xsi:type="IslandLocation">
locations: one for Qi's walnut room, one for the shipwreck, and one
for the mushroom cave.

The fix is to prefer <name>/<Name> elements over xsi:type. This, of
course, results in several new locations that all need to be added to
the locations list.